updade desafios aula 7

+++ desafios/desafio08.py
@@ -6,7 +6,7 @@ print('')
 print('===== DESAFIO 8 =====')
 print('')
 
-me = int(input('Quantos metros: '))
+me = float(input('Quantos metros: '))
 
 print('O valor de {} metro(s) em centimetros é {}.'.format(me, me*100))
 print('O valor de {} metro(s) em milimetros é {}.'.format(me, me*1000))
